feature/VB0469-move-code-from-choise-respose
- Move all code from action controller response types to services/response_base
- Create a new service to implement methods from response base and add response type to include on response base
